[mlir] move LinalgToStandard to Linalg as ConvertToFunctionCalls
The remnants of the ConvertLinalgToStandard pass were still present in the
codebase under this name, years after the Standard dialect was dismantled.
Practically, this pass / pattern set was only performing the rewrite of Linalg
operaitons to function calls. All this makes the existence of the pass highly
confusing.

Move the logic under Linalg/Transforms, similarly to other "lowerings" from
Linalg, e.g., the one to (affine or SCF) loops. Rename ConvertLinalgToStandard
to ConvertLinalgToFunctionCalls. Merge the two relevant test files, ironically,
one of them was called library-calls.mlir. Simplify the code a little.
